+++ rldiff/exception.py
@@ -1,0 +1,2 @@
+def InvalidRenderModeException(Exception):
+    pass