+++ CHANGELOG.md
@@ -1,11 +1,29 @@
 # Change Log
 Maintaining records for changes and the reasoning behind them. Expanding on what features were specifically added, what structural changes have been made, and any fixes to issues that arise in production or tests.
 ___
-## [0.0.1] - 2024-08-13
+## [0.0.3] - 2024-08-15: Evaluation Scripts
+### Added
+Evaluation scripts: [TICKET-11](https://github.com/users/yeabsiramoges/projects/2/views/1?pane=issue&itemId=74875701)
+
+### Changed
+Added evaluation script as well as random action agent for testing and plotting.
+
+### Fixed
+Buggy tests that did not run properly from [0.0.2]
+
+## [0.0.2] - 2024-08-14: Testing Suite
+### Added
+Setting up tests: [TICKET-3](https://github.com/users/yeabsiramoges/projects/2/views/1?pane=issue&itemId=74845920)
+### Changed
+N/A
+### Fixed
+N/A
+
+## [0.0.1] - 2024-08-13: Gymnasium and Project Environment Setup
 
 ### Added
-Set up project: Ticket [YEAB-1](https://github.com/users/yeabsiramoges/projects/2?pane=issue&itemId=74705749)
-Set up gymnasium environment: Ticket [YEAB-2](https://github.com/users/yeabsiramoges/projects/2/views/1?pane=issue&itemId=74710428)
+Set up project: Ticket [TICKET-1](https://github.com/users/yeabsiramoges/projects/2?pane=issue&itemId=74705749)
+Set up gymnasium environment: Ticket [TICKET-2](https://github.com/users/yeabsiramoges/projects/2/views/1?pane=issue&itemId=74710428)
 
 ### Changed
 N/A